move types to top of component directory
