+++ packages/components/src/Combobox/components/Trigger/index.ts
@@ -1,3 +1,3 @@
 export { TriggerButton } from "./TriggerButton";
 export { TriggerChip } from "./TriggerChip";
-export { TriggerProps } from "./types";
+export { TriggerProps } from "../../types";
